apps/api/pyproject.toml
[tool.poetry]
name = "astro-zoom-api"
version = "0.1.0"
description = "FastAPI backend for Astro-Zoom deep image viewer"
authors = ["Astro-Zoom Contributors"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlmodel = "^0.0.14"
pydantic = "^2.5.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
slowapi = "^0.1.9"
shapely = "^2.0.2"
geojson = "^3.1.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"

[tool.poetry.dev-dependencies]
ruff = "^0.1.9"
mypy = "^1.7.1"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

